+++ processor/src/test/resources/expected/com/example/repository/DaggerDisabledRepositoryRepository.java
@@ -26,7 +26,7 @@ import org.realityforge.braincheck.Guards;
 @Generated("arez.processor.ArezProcessor")
 @ArezComponent(
     nameIncludesId = false,
-    dagger = Injectible.FALSE
+    dagger = Injectible.DISABLE
 )
 @Singleton
 public class DaggerDisabledRepositoryRepository {