+++ processor/src/test/resources/input/com/example/inject/MultipleArgsModel.java
@@ -4,7 +4,7 @@ import arez.annotations.Action;
 import arez.annotations.ArezComponent;
 import arez.annotations.Injectible;
 
-@ArezComponent( inject = Injectible.TRUE )
+@ArezComponent( inject = Injectible.ENABLE )
 public class MultipleArgsModel
 {
   public MultipleArgsModel( int i, String foo )