+++ processor/src/test/resources/input/com/example/repository/InjectEnabledRepository.java
@@ -6,7 +6,7 @@ import arez.annotations.Observable;
 import arez.annotations.Repository;
 import javax.annotation.Nonnull;
 
-@Repository( inject = Injectible.TRUE )
+@Repository( inject = Injectible.ENABLE )
 @ArezComponent
 public class InjectEnabledRepository
 {