+++ CHANGELOG.md
@@ -8,6 +8,10 @@
   it introduced.
 * **\[processor\]** Always generate the Dagger2 module as a public interface rather than letting
   the components access modifier dictate the access modifier of the dagger module.
+* 💥 **\[annotations\]** Rename the `Injectible` constants to improve clarity;
+  - `TRUE` renamed to `ENABLED`
+  - `FALSE` renamed to `DISABLED`
+  - `IF_DETECTED` renamed to `AUTODETECT`
 
 ### [v0.39](https://github.com/arez/arez/tree/v0.39) (2018-01-09)
 [Full Changelog](https://github.com/arez/arez/compare/v0.38...v0.39)